+++ src/components/CourseComponents/CreateCourseModal.vue
@@ -3,7 +3,7 @@ import { defineEmits } from "vue/dist/vue";
 import { ref } from "vue";
 import { postCourse } from "@/ts/rest-clients/course-rest-client";
 import { useToast } from "vue-toastification";
-import { ICourse } from "@/ts/modals/models";
+import { ICourse } from "@/ts/models/overworld-models";
 import { defineProps, watch } from "vue";
 import { validateSemester } from "@/ts/validation/validate";
