+++ src/components/EditMinigameModals/EditMinigameConfigurationModal.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { defineProps, defineEmits, ref, watch } from "vue";
-import { ITask } from "@/ts/modals/models";
+import { ITask } from "@/ts/models/overworld-models";
 
 const props = defineProps<{
   minigame: ITask;
