+++ src/components/MapImageModal.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import config from "@/config";
 import { defineProps, defineEmits, ref, watch } from "vue";
-import { MapType } from "@/ts/modals/models";
+import { MapType } from "@/ts/models/overworld-models";
 
 const props = defineProps<{
   worldIndex: string | number;
@@ -78,6 +78,7 @@ function getImage(): string {
     size="xl"
     v-model="showModal"
     @hidden="hiddenModal"
+    v-if="showModal"
   >
     <b-img :src="getImage()" fluid></b-img>
   </b-modal>
