+++ src/ts/models/finitequiz-models.ts
@@ -1,0 +1,19 @@
+export interface IFinitequizConfiguration {
+  id?: string;
+  questions: IFinitequizQuestion[];
+}
+
+export interface IFinitequizQuestion {
+  id?: string;
+  text: string;
+  rightAnswer: string;
+  wrongAnswers: string[];
+}
+
+export class FinitequizConfiguration implements IFinitequizConfiguration {
+  id?: string;
+  questions: IFinitequizQuestion[];
+  public constructor(questions: IFinitequizQuestion[]) {
+    this.questions = questions;
+  }
+}
