+++ src/ts/rest-clients/area-rest-client.ts
@@ -1,6 +1,6 @@
 import { AxiosResponse } from "axios";
 
-import { IArea } from "../modals/models";
+import { IArea } from "../models/overworld-models";
 import { getWorld, updateWorld } from "./world-rest-client";
 import { getDungeon, updateDungeon } from "./dungeon-rest-client";
