+++ tests/unit/CoursesView.spec.ts
@@ -2,8 +2,8 @@ import { mount, VueWrapper, flushPromises } from "@vue/test-utils";
 import mockAxios from "jest-mock-axios";
 import router from "@/router/index";
 import CoursesView from "@/views/CoursesView.vue";
-import { Course, ICourse } from "@/ts/modals/models";
-import BootstrapVue3, { BCol, BRow } from "bootstrap-vue-3";
+import { Course, ICourse } from "@/ts/models/overworld-models";
+import BootstrapVue3, { BCol } from "bootstrap-vue-3";
 import config from "@/config";
 
 jest.mock("axios");
