+++ tests/unit/EditChickenshockConfigurationModal.spec.ts
@@ -1,10 +1,7 @@
 import { flushPromises, mount, VueWrapper } from "@vue/test-utils";
 import EditChickenshockConfigurationModal from "@/components/EditMinigameModals/EditChickenshockConfigurationModal.vue";
-import {
-  IChickenshockConfiguration,
-  ITask,
-  Minigame,
-} from "@/ts/modals/models";
+import { ITask, Minigame } from "@/ts/models/overworld-models";
+import { IChickenshockConfiguration } from "@/ts/models/chickenshock-models";
 import BootstrapVue3, { BButton, BModal } from "bootstrap-vue-3";
 import router from "@/router";
 import mockAxios from "jest-mock-axios";
